requirements.txt
flask>=2.3.0
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0
werkzeug>=2.3.0
mutagen>=1.46.0
imageio-ffmpeg>=0.5.1
pillow>=10.0.0
requests>=2.31.0
